InventoryManagement/requirements.txt
absl-py
annotated-types
anyio
arabic-reshaper
asgiref
asn1crypto
astunparse
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cssselect2
cycler
defusedxml
Django
django-cors-headers
djangorestframework
et_xmlfile
fastapi
filelock
Flask
flatbuffers
fonttools
fpdf2
fsspec
gast
google-pasta
greenlet
grpcio
h5py
html5lib
idna
itsdangerous
Jinja2
joblib
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
mpmath
namex
narwhals
networkx
nltk
numpy
opencv-python
openpyxl
opt_einsum
optree
oscrypto
packaging
pandas
pillow
plotly
protobuf
psycopg2-binary
pycparser
pydantic
pydantic_core
Pygments
pyHanko
pyhanko-certvalidator
pyparsing
pypdf
python-bidi
python-dateutil
pytz
PyYAML
regex
reportlab
requests
rich
scikit-learn
scipy
seaborn
six
sniffio
soupsieve
SQLAlchemy
sqlparse
starlette
svglib
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tinycss2
torch
torchvision
tqdm
typing-inspection
typing_extensions
tzdata
tzlocal
uritools
urllib3
webencodings
Werkzeug
wrapt
xhtml2pdf
xlrd
xlsxwriter
